move batch serialization tests to kernel tests
